% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aovp_bpr.R
\name{aovp_bpr}
\alias{aovp_bpr}
\title{Wrap around lmPerm::aovp()}
\usage{
aovp_bpr(formula, data = NULL)
}
\arguments{
\item{formula}{A valid R formula}

\item{data}{Data (with at least two columns)}
}
\value{
An aovp object
}
\description{
Wrap around lmPerm::aovp()
}
