% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSCC487-package.R
\docType{package}
\name{DSCC487-package}
\alias{DSCC487}
\alias{DSCC487-package}
\title{DSCC487: Datasets and miscellaneous functions for DSCC 487}
\description{
This project contains miscellanseous functions and several datasets to be used in DSCC 487.
}
\author{
\strong{Maintainer}: Barney Ricca \email{barneyricca@gmail.com}

}
\keyword{internal}
