% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquitoes.R
\docType{data}
\name{mosquitoes_df}
\alias{mosquitoes_df}
\title{mosquitoes_df}
\format{
An object of class \code{data.frame} with 43 rows and 2 columns.
}
\source{
\url{https://www.routledge.com/Nonlinear-Dynamical-Systems-Analysis-for-the-Behavioral-Sciences-Using-Real-Data/Guastello-Gregson/p/book/9781439819975?srsltid=AfmBOooMo5hYC4OG-x8bUNkPVhBbLo-IyQJ-XCyF-5YG8HDlxT238AvN}
}
\usage{
mosquitoes_df
}
\description{
Data from somewhere
}
\author{
Barney Ricca barneyricca@gmail.com
}
\keyword{mosquitoes}
